+++ project/src/index.tsx
@@ -3,13 +3,14 @@ import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
 import {Provider} from 'react-redux';
 import {store} from './store';
-import {fetchFilmsAction, fetchPromoFilmAction, checkAuthAction} from './store/api-actions';
+import {fetchFilmsAction, fetchPromoFilmAction, checkAuthAction, fetchFavoriteFilmsAction} from './store/api-actions';
 import {ToastContainer} from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 store.dispatch(fetchFilmsAction());
 store.dispatch(fetchPromoFilmAction());
 store.dispatch(checkAuthAction());
+store.dispatch(fetchFavoriteFilmsAction());
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement,
