--- project/src/mocks/comments.ts
@@ -1,64 +1,0 @@
-import {Comment } from '../types/types';
-
-export const comments: Comment[] = [
-  {
-    comment: 'Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director\'s funniest and most exquisitely designed movies in years.',
-    date: 'Wed Aug 03 2022 13:12:37 GMT+0300 (Москва, стандартное время)',
-    id: 1,
-    rating: 8.9,
-    user: {
-      id: 4,
-      name: 'Kate Muir'
-    }
-  },
-  {
-    comment: 'Anderson\'s films are too precious for some, but for those of us willing to lose ourselves in them, they\'re a delight. "The Grand Budapest Hotel" is no different, except that he has added a hint of gravitas to the mix, improving the recipe.',
-    date: 'Wed Aug 18 2015 13:12:37 GMT+0300 (Москва, стандартное время)',
-    id: 2,
-    rating: 8.0,
-    user: {
-      id: 2,
-      name: 'Bill Goodykoontz'
-    }
-  },
-  {
-    comment: 'I didn\'t find it amusing, and while I can appreciate the creativity, it\'s an hour and 40 minutes I wish I could take back.',
-    date: 'Wed Aug 13 2012 13:12:37 GMT+0300 (Москва, стандартное время)',
-    id: 3,
-    rating: 8.0,
-    user: {
-      id: 1,
-      name: 'Amanda Greever'
-    }
-  },
-  {
-    comment: 'The mannered, madcap proceedings are often delightful, occasionally silly, and here and there, gruesome and/or heartbreaking.',
-    date: 'Wed Aug 03 2022 13:12:37 GMT+0300 (Москва, стандартное время)',
-    id: 4,
-    rating: 7.2,
-    user: {
-      id: 3,
-      name: 'Matthew Lickona'
-    }
-  },
-  {
-    comment: 'It is certainly a magical and childlike way of storytelling, even if the content is a little more adult.',
-    date: 'Wed Aug 03 2022 13:12:37 GMT+0300 (Москва, стандартное время)',
-    id: 5,
-    rating: 7.6,
-    user: {
-      id: 1,
-      name: 'Paula Fleri-Soler'
-    }
-  },
-  {
-    comment: 'It is certainly a magical and childlike way of storytelling, even if the content is a little more adult.',
-    date: 'Wed Aug 03 2022 13:12:37 GMT+0300 (Москва, стандартное время)',
-    id: 6,
-    rating: 7.0,
-    user: {
-      id: 7,
-      name: 'Paula Fleri-Soler'
-    }
-  }
-];
