+++ project/src/pages/movie-page-screen/movie-page-screen.tsx
@@ -16,26 +16,18 @@ import UserBlockAnonim from '../../components/user-block-anonim/user-block-anoni
 import UserBlockAuth from '../../components/user-block-auth/user-block-auth';
 import {getFilm, getSimilarFilms} from '../../store/film-process/selectors';
 import {getAuthorizationStatus, getuserData} from '../../store/user-process/selectors';
-import {setFavoriteFilmAction} from '../../store/api-actions';
+import MyListButton from '../../components/my-list-button/my-list-button';
 
 function MoviePageScreen(): JSX.Element {
   const film = useAppSelector(getFilm);
   const similarFilms = useAppSelector(getSimilarFilms);
   const userData = useAppSelector(getuserData);
   const authorizationStatus = useAppSelector(getAuthorizationStatus);
+
   const dispatch = useAppDispatch();
   const navigate = useNavigate();
   const params = useParams();
   const filmId = Number(params?.id);
-
-  const favoriteData = {
-    filmId: film?.id,
-    status: film?.isFavorite ? 0 : 1
-  };
-
-  const handleFavoriteButtonClick = () => {
-    dispatch(setFavoriteFilmAction(favoriteData));
-  };
 
   useEffect(() => {
     dispatch(fetchFilmAction(filmId));
@@ -78,13 +70,7 @@ function MoviePageScreen(): JSX.Element {
                   </svg>
                   <span>Play</span>
                 </button>
-                <button className="btn btn--list film-card__button" type="button" onClick={handleFavoriteButtonClick}>
-                  <svg viewBox="0 0 19 20" width="19" height="20">
-                    <use xlinkHref="#add"></use>
-                  </svg>
-                  <span>My list</span>
-                  <span className="film-card__count">9</span>
-                </button>
+                <MyListButton film={film} />
                 {authorizationStatus === AuthorizationStatus.Auth
                 &&
                 <Link to={AppRoute.AddReview} className="btn film-card__button">Add review</Link>}
